Skyline: Move some remaining UI code out of the Model folder
