+++ pwiz_tools/Skyline/EditUI/AddPeakCompareDlg.cs
@@ -24,6 +24,7 @@ using System.Linq;
 using System.Windows.Forms;
 using pwiz.Skyline.Alerts;
 using pwiz.Skyline.Controls;
+using pwiz.Skyline.FileUI;
 using pwiz.Skyline.Model;
 using pwiz.Skyline.Model.Results.Scoring;
 using pwiz.Skyline.Properties;
